requirements.txt
flask
flask-cors
python-dotenv
groq
pymupdf
mysql-connector-python
werkzeug
requests
